Switched from render hook to class extension
