+++ src/module/applications/_module.mjs
@@ -1,6 +1,7 @@
 export * as sidebar from "./sidebar/_module.mjs";
 export * as apps from "./apps/_module.mjs";
 export * as hooks from "./hooks/_module.mjs";
+export * as hud from "./hud/_module.mjs";
 export * as sheets from "./sheets/_module.mjs";
 
 import * as elements from "./elements/_module.mjs";
