+++ src/module/applications/hooks/_module.mjs
@@ -1,4 +1,3 @@
 export * from "./chatMessageHTML.mjs";
 export * from "./combatantConfig.mjs";
-export * from "./token-hud.mjs";
 export * from "./tokenApplication.mjs";
