+++ src/module/applications/hud/_module.mjs
@@ -1,0 +1,1 @@
+export { default as DrawSteelTokenHUD } from "./token-hud.mjs";
